+++ library/src/index.ts
@@ -3,7 +3,7 @@ import AsyncApiComponentWP from './containers/AsyncApi/Standalone';
 
 export { AsyncApiProps } from './containers/AsyncApi/AsyncApi';
 export { ConfigInterface } from './config/config';
-export { FetchingSchemaInterface } from './types';
+export { FetchingSchemaInterface, ExtensionComponentProps } from './types';
 
 import { hljs } from './helpers';
